voice-agents/interrupt_control/pyproject.toml
[project]
name = "atoms-interrupt-control"
version = "0.1.0"
description = "Runtime configuration changes for voice agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "smallestai>=4.2.0",
    "python-dotenv",
    "loguru",
]
